requirements.txt
colorama==0.4.4
gitdb==4.0.7
GitPython==3.1.17
PyYAML==5.4.1
smmap==4.0.0
typing-extensions==3.10.0.0
